requirements/test.txt
case>=1.3.1
pytest>=6.0.2
pytest-django
pytest-benchmark
pytz>dev
psycopg2cffi
